ExTumblr v0.0.1 ExTumblr.Avatar

Provide a function to request the avatar url of a blog.

Types

avatar_size ::
  16 |
  24 |
  30 |
  40 |
  48 |
  64 |
  96 |
  128 |
  512

Functions

request(client, blog_identifier, size \\ 64)

Specs

request(ExTumblr.Client.t, String.t, avatar_size) :: String.t

Request the url of the avatar of a blog.

Official documentation